% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arrayEx}
\alias{arrayEx}
\title{Simple MsgPack Example}
\usage{
arrayEx()
}
\value{
A boolean value of \code{TRUE} is returned, but the function exists
for its side effect.
}
\description{
Simple MsgPack Example
}
\details{
The function provides a simple illustration of MessagePack.
}
\seealso{
The MessagePack documentation, particularly the msgpack-c examples.
}
