% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{enumEx}
\alias{enumEx}
\title{Second simple MsgPack Example}
\usage{
enumEx()
}
\value{
A boolean value of \code{TRUE} is returned, but the function exists
for its side effect.
}
\description{
Second simple MsgPack Example
}
\details{
The function provides a simple illustration of MessagePack.
}
\seealso{
The MessagePack documentation, particularly the msgpack-c examples.
}

