% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{msgpack_format}
\alias{msgpack_format}
\title{Format data for 'MsgPack'}
\usage{
msgpack_format(x)
}
\arguments{
\item{x}{An r object.}
}
\value{
A formatted R object to use as input to msgpack_pack.  Vectors are converted into Lists.
}
\description{
A helper function to format R data for input to 'MsgPack'.
}
\examples{
library(RcppMsgPack)
x <- msgpack_format(1:10)
x_packed <- msgpack_pack(x)
x_unpacked <- msgpack_unpack(x_packed)
x_simplified <- msgpack_simplify(x_unpacked)
}
