% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{msgpack_simplify}
\alias{msgpack_simplify}
\title{Simplify 'MsgPack'}
\usage{
msgpack_simplify(x)
}
\arguments{
\item{x}{Return object from msgpack_unpack.}
}
\value{
A simplified return object from msgpack_unpack.  Lists of all the same type are concatenated into an atomic vector.  Maps are simplified to named lists or named vectors as appropriate.  NULLs are converted to NAs if simplified to vector.
}
\description{
A helper function for simplifying a 'MsgPack' return object.
}
\examples{
library(RcppMsgPack)
x <- msgpack_format(1:10)
x_packed <- msgpack_pack(x)
x_unpacked <- msgpack_unpack(x_packed)
x_simplified <- msgpack_simplify(x_unpacked)
}
