\name{RcppMsgPack-package}
\alias{RcppMsgPack-package}
\alias{RcppMsgPack}
\docType{package}
\title{\packageTitle{RcppMsgPack}}
\description{\packageDescription{RcppMsgPack}}
\section{Package Content}{\packageIndices{RcppMsgPack}}
\author{\packageAuthor{RcppMsgPack}}
\section{Maintainer}{\packageMaintainer{RcppMsgPack}}
\keyword{ package }
