\name{RcppMsgPack-package}
\alias{RcppMsgPack-package}
\alias{RcppMsgPack}
\docType{package}
\title{
  Headers for the MessagePack Binary Serialization Library
}
\description{
 'MessagePack' is an efficient binary serialization format.
  It lets you exchange data among multiple languages like JSON. But it's
  faster and smaller. Small integers are encoded into a single byte, and
  typical short strings require only one extra byte in addition to the strings
  themselves.
}
\details{
 'MessagePack' is an efficient binary serialization format.
  It lets you exchange data among multiple languages like JSON. But it's
  faster and smaller. Small integers are encoded into a single byte, and
  typical short strings require only one extra byte in addition to the strings
  themselves.
  
  Bug reports can also be registered at the GitHub issue tracker at
  \url{https://github.com/eddelbuettel/rcppmsgpack/issues}.
}
\author{
  Dirk Eddelbuettel

  Maintainer: Dirk Eddelbuettel <edd@debian.org>
}
\references{
  http://msgpack.org/
}
\keyword{ package }
\examples{
  # None
}
